sm8550-common: Downgrade to ARMv8.2-A
Clang assumes that ARMv9-A chips always implement SVE2,
but SM8550 (Snapdragon 8 Gen 2) does not, so it broke when
aosp/2292699 made the "armv9-a" variant actually set ARMv9-A.
Before, it was actually using ARMv8.2-A with some additional
flags due to a variety of SVE-specific toolchain and
configuration issues that are fixed now.

Downgrade SM8550 to the "ARMv8.2-A with dotprod extension"
arch variant, which is the closest to the previous
configuration, until we can add a proper "ARMv9-A without SVE"
arch variant.

Bug: 364383722
Test: presubmit
Change-Id: I69578aad149f6a28955031ca24bc1d0cc7cc0388